backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
argostranslate==1.11.0
asgiref==3.11.1
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
ctranslate2==4.7.1
cymem==2.0.13
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.0
emoji==2.15.0
ffmpeg-python==0.2.0
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.2.0
future==1.0.0
gTTS==2.5.4
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
minisbd==0.9.5
more-itertools==10.8.0
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
onnxruntime==1.24.4
openai-whisper==20250625
packaging==26.0
preshed==3.0.13
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyobjc==12.1
pyobjc-core==12.1
pyobjc-framework-Accessibility==12.1
pyobjc-framework-Accounts==12.1
pyobjc-framework-AddressBook==12.1
pyobjc-framework-AdServices==12.1
pyobjc-framework-AdSupport==12.1
pyobjc-framework-AppleScriptKit==12.1
pyobjc-framework-AppleScriptObjC==12.1
pyobjc-framework-ApplicationServices==12.1
pyobjc-framework-AppTrackingTransparency==12.1
pyobjc-framework-ARKit==12.1
pyobjc-framework-AudioVideoBridging==12.1
pyobjc-framework-AuthenticationServices==12.1
pyobjc-framework-AutomaticAssessmentConfiguration==12.1
pyobjc-framework-Automator==12.1
pyobjc-framework-AVFoundation==12.1
pyobjc-framework-AVKit==12.1
pyobjc-framework-AVRouting==12.1
pyobjc-framework-BackgroundAssets==12.1
pyobjc-framework-BrowserEngineKit==12.1
pyobjc-framework-BusinessChat==12.1
pyobjc-framework-CalendarStore==12.1
pyobjc-framework-CallKit==12.1
pyobjc-framework-Carbon==12.1
pyobjc-framework-CFNetwork==12.1
pyobjc-framework-Cinematic==12.1
pyobjc-framework-ClassKit==12.1
pyobjc-framework-CloudKit==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Collaboration==12.1
pyobjc-framework-ColorSync==12.1
pyobjc-framework-CompositorServices==12.1
pyobjc-framework-Contacts==12.1
pyobjc-framework-ContactsUI==12.1
pyobjc-framework-CoreAudio==12.1
pyobjc-framework-CoreAudioKit==12.1
pyobjc-framework-CoreBluetooth==12.1
pyobjc-framework-CoreData==12.1
pyobjc-framework-CoreHaptics==12.1
pyobjc-framework-CoreLocation==12.1
pyobjc-framework-CoreMedia==12.1
pyobjc-framework-CoreMediaIO==12.1
pyobjc-framework-CoreMIDI==12.1
pyobjc-framework-CoreML==12.1
pyobjc-framework-CoreMotion==12.1
pyobjc-framework-CoreServices==12.1
pyobjc-framework-CoreSpotlight==12.1
pyobjc-framework-CoreText==12.1
pyobjc-framework-CoreWLAN==12.1
pyobjc-framework-CryptoTokenKit==12.1
pyobjc-framework-DataDetection==12.1
pyobjc-framework-DeviceCheck==12.1
pyobjc-framework-DeviceDiscoveryExtension==12.1
pyobjc-framework-DictionaryServices==12.1
pyobjc-framework-DiscRecording==12.1
pyobjc-framework-DiscRecordingUI==12.1
pyobjc-framework-DiskArbitration==12.1
pyobjc-framework-DVDPlayback==12.1
pyobjc-framework-EventKit==12.1
pyobjc-framework-ExceptionHandling==12.1
pyobjc-framework-ExecutionPolicy==12.1
pyobjc-framework-ExtensionKit==12.1
pyobjc-framework-ExternalAccessory==12.1
pyobjc-framework-FileProvider==12.1
pyobjc-framework-FileProviderUI==12.1
pyobjc-framework-FinderSync==12.1
pyobjc-framework-FSEvents==12.1
pyobjc-framework-FSKit==12.1
pyobjc-framework-GameCenter==12.1
pyobjc-framework-GameController==12.1
pyobjc-framework-GameKit==12.1
pyobjc-framework-GameplayKit==12.1
pyobjc-framework-GameSave==12.1
pyobjc-framework-HealthKit==12.1
pyobjc-framework-ImageCaptureCore==12.1
pyobjc-framework-InputMethodKit==12.1
pyobjc-framework-InstallerPlugins==12.1
pyobjc-framework-InstantMessage==12.1
pyobjc-framework-Intents==12.1
pyobjc-framework-IntentsUI==12.1
pyobjc-framework-IOBluetooth==12.1
pyobjc-framework-IOBluetoothUI==12.1
pyobjc-framework-IOSurface==12.1
pyobjc-framework-iTunesLibrary==12.1
pyobjc-framework-KernelManagement==12.1
pyobjc-framework-LatentSemanticMapping==12.1
pyobjc-framework-LaunchServices==12.1
pyobjc-framework-libdispatch==12.1
pyobjc-framework-libxpc==12.1
pyobjc-framework-LinkPresentation==12.1
pyobjc-framework-LocalAuthentication==12.1
pyobjc-framework-LocalAuthenticationEmbeddedUI==12.1
pyobjc-framework-MailKit==12.1
pyobjc-framework-MapKit==12.1
pyobjc-framework-MediaAccessibility==12.1
pyobjc-framework-MediaExtension==12.1
pyobjc-framework-MediaLibrary==12.1
pyobjc-framework-MediaPlayer==12.1
pyobjc-framework-MediaToolbox==12.1
pyobjc-framework-Metal==12.1
pyobjc-framework-MetalFX==12.1
pyobjc-framework-MetalKit==12.1
pyobjc-framework-MetalPerformanceShaders==12.1
pyobjc-framework-MetalPerformanceShadersGraph==12.1
pyobjc-framework-MetricKit==12.1
pyobjc-framework-MLCompute==12.1
pyobjc-framework-ModelIO==12.1
pyobjc-framework-MultipeerConnectivity==12.1
pyobjc-framework-NaturalLanguage==12.1
pyobjc-framework-NetFS==12.1
pyobjc-framework-Network==12.1
pyobjc-framework-NetworkExtension==12.1
pyobjc-framework-NotificationCenter==12.1
pyobjc-framework-OpenDirectory==12.1
pyobjc-framework-OSAKit==12.1
pyobjc-framework-OSLog==12.1
pyobjc-framework-PassKit==12.1
pyobjc-framework-PencilKit==12.1
pyobjc-framework-PHASE==12.1
pyobjc-framework-Photos==12.1
pyobjc-framework-PhotosUI==12.1
pyobjc-framework-PreferencePanes==12.1
pyobjc-framework-PushKit==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-QuickLookThumbnailing==12.1
pyobjc-framework-ReplayKit==12.1
pyobjc-framework-SafariServices==12.1
pyobjc-framework-SafetyKit==12.1
pyobjc-framework-SceneKit==12.1
pyobjc-framework-ScreenCaptureKit==12.1
pyobjc-framework-ScreenSaver==12.1
pyobjc-framework-ScreenTime==12.1
pyobjc-framework-ScriptingBridge==12.1
pyobjc-framework-SearchKit==12.1
pyobjc-framework-Security==12.1
pyobjc-framework-SecurityFoundation==12.1
pyobjc-framework-SecurityInterface==12.1
pyobjc-framework-SecurityUI==12.1
pyobjc-framework-SensitiveContentAnalysis==12.1
pyobjc-framework-ServiceManagement==12.1
pyobjc-framework-SharedWithYou==12.1
pyobjc-framework-SharedWithYouCore==12.1
pyobjc-framework-ShazamKit==12.1
pyobjc-framework-Social==12.1
pyobjc-framework-SoundAnalysis==12.1
pyobjc-framework-Speech==12.1
pyobjc-framework-SpriteKit==12.1
pyobjc-framework-StoreKit==12.1
pyobjc-framework-Symbols==12.1
pyobjc-framework-SyncServices==12.1
pyobjc-framework-SystemConfiguration==12.1
pyobjc-framework-SystemExtensions==12.1
pyobjc-framework-ThreadNetwork==12.1
pyobjc-framework-UniformTypeIdentifiers==12.1
pyobjc-framework-UserNotifications==12.1
pyobjc-framework-UserNotificationsUI==12.1
pyobjc-framework-VideoSubscriberAccount==12.1
pyobjc-framework-VideoToolbox==12.1
pyobjc-framework-Virtualization==12.1
pyobjc-framework-Vision==12.1
pyobjc-framework-WebKit==12.1
pyttsx3==2.99
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
sacremoses==0.1.1
sentencepiece==0.2.1
shellingham==1.5.4
smart_open==7.5.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.5
srsly==2.5.3
stanza==1.10.1
sympy==1.14.0
thinc==8.3.11
tiktoken==0.12.0
torch==2.10.0
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
wrapt==2.1.2